Move code out of retroarch.c and into input_driver.c
